Add presentation initial style [See details]
Centralize buttun and 'div' text.

The 'z-index' property in conjunction of 'position' is set so that the text in this section is positioned in front of the background image.

Warning: this commit does not applies the desired look and feel of the
background image: it needs to be darkened a little bit and it needs to grow
horizontally following the window width.