Add initial style of navbar for mobile devices [See more]
This commit just adds the logo icon to the navbar, and fixes the navbar position on
the context of a mobile page.

It's still needed to add a hamburger icon in the navbar. The hamburger
must be in the right side. This implies that the logo icon must be
placed in the left side of the navbar.