Add styles of nav links for mobile; add style in texts classes [See more]
The `display: inline-block` in the nav's elements,
along with `text-align: justify` in the nav itself
allow us to achieve the desired style: Aria logo and menu icon
positioned in the left and right, respectively, and the nav content
links behind those.

The `text-decoration: none` is necessary to remove underline when the
classes `.small-text` and `.small-text-uppercase` are applied to <a> elements.