Fix horizontal disposition of `li`s in intro steps [See more]
The max-width along with the padding is set in the `li`s to prevent the text
in it of being too close of the adjacents `li`s.

This commit increases the spacing between adjacent `li`s in the intro
steps section, so that those items are centered on medium and large screens.