Large refactor on texts elements [See more]
This commit refactores the text into atomic pieces, so that it can be
reused.

This commit adds the class `text-container` to the `div` that wraps all
texts in `intro__content` in order to ajust the width and the bottom
margin of it.

On the nested `text-container`, the `p` tag that wraps the `q` tag is
set so that the bottom margin works (the class `text--distance-small`
does it).

The `text-header` block, along with its elements (`text-header__title`
and `text-header__message` encapsulates the header style that is going to be
used more on the site.