Add teams initial markup and style
The 'flex' block is to generalize a responsive flex container.

This commit hardcodes the 'image-container' size in the team section in a way that keeps the aspect ratio of the team members' photo.

Positionates the image element of 'image-container' in the team section as absolutly relative to the 'image-container'. This is done to ensure that the sizes of those will be equal when the size of the 'image-container' is specified. (Thoses sizes need to be equal, otherwise the transition zoom effect on hover won't behave correctly).