Add light theme style in form item; add style for textarea [See more]
This commit replaces SASS variables to CSS variables in form-item in
order to customize the theme of the form by changing those variables in
the class modifier 'form-item--theme-light'.

Also, this commit sets the color property in the form-item itself, to allow
child elements of the form-item to get that color with 'color: inherit'.
This allow us to change the color of all text elements inside a
form-item by just changing the color of that form-item.

Finally, this commit fixes the transition effect of the
'form-item__placeholder'. It reduces the font-size a little to preserve the placeholder's left-margin when the transition is fired.