Fix devServer configuration [See more]
The auto-reload now works when we change some code. The problem was that 'index.html' is requesting 'main.css' and 'main.js' in 'dist' directory. By default, devServer output bundles to 'localhost:8080/', which is different than the path that our html file is expecting - 'localhost:8080/dist/'. Thus, to send generated bundles to 'dist', this commit sets the 'publicPath' property in devServer config file to '/dist/'.