Make devServer refresh browser on index.html change [See more]
This commit removes the 'contentBase' property in the devServer
configuration to make it use the default value, which is the root of the
project. As 'index.html' is placed on this directory, devServer will
serve it as a static file, and devServer won't live-reload on changes to
it. That's why this commit sets 'watchContentBase' to true, so that
devServer will reload on 'index.html' changes.