Fix svg in menu icon position by centralizing it on a wrapper [See more]
The close icon was a little bit to the right of the hamburger menu icon.
So to fix this, this commit centralizes thoses icons inside a wrapper
div element.

Also, this commit applies 'display: inline-block' to immediate children of the nav element. This makes obsolete the prior positioning hacks on the 'nav__menu-icon' (e.g.: 'right: 6%'), as 'text-align: justify' set on 'nav' class ensures the desired positions of the logo and of the menu icon.