Fix transition fade out and fade in effect of tab body [See more]
Replaces 'height' with 'max-height', because only the latter can be transitioned over.

In the fade out effect, the transition over 'opacity' begins immediatly and, when it finishes, the 'max-height' property is set to 0 right away, in order to the next effect (fade in) occours in the correct position.

In the fade in effect, both transitions delays until the fade in transition finishes. And when this happens, the 'max-height' is adjusted immediatly, to show the selected tab body element being faded in due to it's transition over 'opacity'.